% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc}
\alias{calc}
\title{Calc}
\usage{
calc(p1, q1, p2, q2)
}
\arguments{
\item{p1}{preço produto 1}

\item{q1}{quantidade produto 1}

\item{p2}{preço produto 2}

\item{q2}{quantidade produto 2}
}
\description{
Calc
}
